requirements.txt
argcomplete==1.12.0
asgiref==3.4.1
backcall==0.1.0
Beaker==1.10.0
beautifulsoup4==4.9.3
blivet==3.3.3
blivet-gui==2.3.0
Brlapi==0.8.2
cffi==1.14.5
chardet==4.0.0
chrome-gnome-shell==0.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
cupshelpers==1.0
dasbus==1.4
dbus-python==1.2.18
decorator==4.4.2
defusedxml==0.7.1
distro==1.5.0
Django==3.2.8
django-cors-headers==3.10.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
fros==1.1
gpg==1.15.1
humanize==0.5.1
idna==2.10
ipython==7.20.0
ipython-genutils==0.1.0
itypes==1.2.0
jedi==0.17.2
Jinja2==3.0.2
langtable==0.0.56
libcomps==0.1.18
lxml==4.6.3
Mako==1.1.4.dev0
MarkupSafe==2.0.1
nftables==0.1
ntplib==0.3.3
oauthlib==3.1.1
olefile==0.46
ordered-set==4.0.1
packaging==20.9
parso==0.8.0
Paste==3.5.0
pexpect==4.8.0
pickleshare==0.7.5
pid==2.2.3
Pillow==8.1.2
ply==3.11
productmd==1.33
prompt-toolkit==3.0.5
psycopg2-binary==2.9.1
ptyprocess==0.6.0
pwquality==1.4.4
pycairo==1.20.0
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.1
pycurl==7.43.0.6
pyenchant==3.2.1
Pygments==2.7.4
PyGObject==3.40.1
pyinotify==0.9.6
PyJWT==2.3.0
pykickstart==3.32
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyparted==3.11.7
PySocks==1.7.1
python-augeas==0.5.0
python-dateutil==2.8.1
python-meh==0.50
python3-openid==3.2.0
pytz==2021.3
pyudev==0.22.0
pyxdg==0.27
requests==2.25.1
requests-file==1.5.1
requests-ftp==0.3.1
requests-oauthlib==1.3.0
rpm==4.16.1.3
selinux==3.2
sepolicy==3.2
setools==4.4.0
simpleaudio==1.0.4
simplegeneric==0.8.1
simpleline==1.8
six==1.15.0
slip==0.6.4
slip.dbus==0.6.4
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sos==4.1
soupsieve==2.2
sqlparse==0.4.2
systemd-python==234
Tempita==0.5.1
traitlets==5.0.5
uritemplate==4.1.1
urllib3==1.25.10
wcwidth==0.2.5
